// src/app/models/DatesDashboard.ts
export class DatesDashboard{
  entryDate?: Date;
  lunchTimeDate?: Date;
  returnLunchTimeDate?: Date;
  departureTimeDate?: Date;
  workedHours?: number;
  collaboratorId?: number;
  entryHours?: number;

  entry?: string;
  lunchTime?: string;
  returnLunchTime?: string;
  departureTime?: string;
}